Making different types of project

+++ index.js
@@ -16,17 +16,19 @@ async function askType() {
     type: 'list',
     message: 'Select type of the project',
     choices: [
-      'Bare project - basically only index.js',
-      'Sample routes - with mainRouter',
-      'Sequelize sample - with sequelize and postgresql project',
+      '1) Bare project - basically only server.js',
+      '2) Sequelize sample - with sequelize and postgresql project',
+      '3) Full setup - sequelize, admin.js, swagger',
+      '4) Custom setup',
     ],
   });
 
   return type.selectType;
 }
 
-const setUpFolders = async () => {
-  console.clear();
+const setUpFolders = async (type = 0) => {
+  // console.clear();
+
   console.log(chalk.blue('Setting up folders'));
   // create folders structure
   fs.mkdirSync(`${dir}/Routes`);
@@ -53,16 +55,18 @@ const finishedMessage = () => {
   );
 };
 
-const main = async () => {
+const main = async (type) => {
+  console.log(type[0]);
   fs.mkdirSync(dir);
-  console.clear();
   // init npm, install modules, git init
-  const spinner = createSpinner('Initializing repository with npm').start();
-  exec(
-    `clear && cd ${process.argv[2]} && npm init -y  && npm i express nodemon bcryptjs express-validator jsonwebtoken`,
-    (error, stdout, stderr) => {
+  const spinner = createSpinner('Initializing repository').start();
+  if (type[0] == 1) {
+    // console.log(1);
+
+    // console.log('Bare');
+    exec(`cd ${process.argv[2]} && npm init -y `, (error, stdout, stderr) => {
       if (error) {
-        spinner.error({ text: `Error` });
+        spinner.error({ text: `Error ${error}` });
         fs.unlinkSync(`${dir}/package.json`);
         fs.rmdirSync(dir);
         process.exit(1);
@@ -73,17 +77,84 @@ const main = async () => {
         fs.rmdirSync(dir);
         process.exit(1);
       }
-      // console.log(`stdout: ${stdout}`);
       spinner.success({ text: 'Finished' });
       setUpFolders();
-    },
-  );
+    });
+  } else if (type[0] == 2) {
+    // console.log(2);
+    exec(
+      `cd ${process.argv[2]} && npm init -y &&  && npm i express nodemon bcryptjs express-validator jsonwebtoken sequelize pg pg-hstore && npm install --save-dev sequelize-cli && npx sequelize-cli init `,
+      (error, stdout, stderr) => {
+        if (error) {
+          spinner.error({ text: `Error` });
+          fs.unlinkSync(`${dir}/package.json`);
+          fs.rmdirSync(dir);
+          process.exit(1);
+        }
+        if (stderr) {
+          spinner.error({ text: `Error` });
+          fs.unlinkSync(`${dir}/package.json`);
+          fs.rmdirSync(dir);
+          process.exit(1);
+        }
+        spinner.success({ text: 'Finished' });
+        setUpFolders();
+      },
+    );
+  } else if (type[0] == 3) {
+    // console.log(3);
+
+    exec(
+      `cd ${process.argv[2]} && npm init -y  && npm i express nodemon bcryptjs express-validator jsonwebtoken sequelize pg pg-hstore swagger-ui-express adminjs @adminjs/sequelize @adminjs/express && npm install --save-dev sequelize-cli`,
+      (error, stdout, stderr) => {
+        if (error) {
+          spinner.error({ text: `Error ${error}` });
+          fs.unlinkSync(`${dir}/package.json`);
+          fs.rmdirSync(dir);
+          process.exit(1);
+        }
+        if (stderr) {
+          spinner.error({ text: `StdError ${error}` });
+          fs.unlinkSync(`${dir}/package.json`);
+          fs.rmdirSync(dir);
+          process.exit(1);
+        }
+        spinner.success({ text: 'Finished' });
+        setUpFolders();
+      },
+    );
+  } else if (type[0] == 4) {
+    // console.log(4);
+
+    console.log('Custom');
+  }
+
+  // exec(
+  //   `clear && cd ${process.argv[2]} && npm init -y  && npm i express nodemon bcryptjs express-validator jsonwebtoken`,
+  //   (error, stdout, stderr) => {
+  //     if (error) {
+  //       spinner.error({ text: `Error` });
+  //       fs.unlinkSync(`${dir}/package.json`);
+  //       fs.rmdirSync(dir);
+  //       process.exit(1);
+  //     }
+  //     if (stderr) {
+  //       spinner.error({ text: `Error` });
+  //       fs.unlinkSync(`${dir}/package.json`);
+  //       fs.rmdirSync(dir);
+  //       process.exit(1);
+  //     }
+  //     // console.log(`stdout: ${stdout}`);
+  //     spinner.success({ text: 'Finished' });
+  //     setUpFolders();
+  //   },
+  // );
 };
 
 if (!fs.existsSync(dir) && process.argv[2] !== undefined) {
   let type = await askType();
   console.log(chalk.blue(type));
-  main();
+  main(type);
 } else {
   console.log('Folder with that name already exists');
 }
